+++ js/string.js
@@ -30,8 +30,9 @@ String.prototype.toTitleCase || Object.defineProperties(String.prototype, {
 	toTitleCase: {
 		value: function () { // preserves `this`
 			return (this.length > 3 && this.isUpperCase() ? this.toLowerCase() : this)
-				.replace(/\w\S*/g, word => word.charAt(0).toUpperCase() + word.substr(1))
-				.trim().replace(/(\s)\s+/g, '$1');
+				.replace(/\w\S*/g, match => match.charAt(0).toUpperCase() + match.substr(1))
+				.replaceAll('  ', ' ')
+				.trimAll();
 		}
 	}
 });
@@ -43,8 +44,22 @@ String.prototype.toSentenceCase || Object.defineProperties(String.prototype, {
 	toSentenceCase: {
 		value: function () { // preserves `this`
 			return (this.length > 3 && this.isUpperCase() ? this.toLowerCase() : this)
-				.replace(/(^\s*\w{1}|\.\s+\w{1})/gm, letter => letter.toUpperCase())
-				.trim().replace(/(\s)\s+/g, '$1');
+				.replace(/(^\s*\w{1}|[.!?]\s+\w{1})/gm, match => match.toUpperCase())
+				.replaceAll('  ', ' ')
+				.trimAll();
+		}
+	}
+});
+
+/**
+ * Returns a string with leading and trailing white space and line terminator characters removed from each line.
+ */
+String.prototype.trimAll || Object.defineProperties(String.prototype, {
+	trimAll: {
+		value: function () { // preserves `this`
+			return this
+				.replace(/(^ +| +$)/gm, '')
+				.trim();
 		}
 	}
 });
@@ -57,8 +72,8 @@ String.prototype.truncate || Object.defineProperties(String.prototype, {
 	truncate: {
 		value: function (max) { // preserves `this`
 			if (max > 0 && this.length > max) {
-				const text = this.stripTags().substr(0, max),
-					last = text.search(/[^\w]+(?:.(?![^\w]))+$/gi);
+				const text = this.toPlainText().substr(0, max),
+					last = text.search(/[^\w]+(?:.(?![^\w]))+$/g);
 				return text.substr(0, last > 0 ? last : max - 1) + '…';
 			}
 			else
@@ -68,12 +83,13 @@ String.prototype.truncate || Object.defineProperties(String.prototype, {
 });
 
 /**
- * Returns an HTML string encoded to prevent cross-site scripting (XSS) attacks.
+ * Returns a string converted into an HTML-encoded string for HTTP transmission.
  */
 String.prototype.htmlEncode || Object.defineProperties(String.prototype, {
 	htmlEncode: {
 		value: function () { // preserves `this`
-			return this.replaceAll('&', '&amp;')
+			return this
+				.replaceAll('&', '&amp;')
 				.replaceAll('<', '&lt;')
 				.replaceAll('>', '&gt;')
 				.replaceAll('"', '&quot;')
@@ -83,12 +99,38 @@ String.prototype.htmlEncode || Object.defineProperties(String.prototype, {
 });
 
 /**
- * Returns an HTML string in plain text, with all HTML tags removed.
+ * Returns a string that has been HTML-encoded converted into a decoded string.
  */
-String.prototype.stripTags || Object.defineProperties(String.prototype, {
-	stripTags: {
+String.prototype.htmlDecode || Object.defineProperties(String.prototype, {
+	htmlDecode: {
 		value: function () { // preserves `this`
-			return this.replace(/<[^>]+>/gi, '');
+			return new DOMParser()
+				.parseFromString(this, 'text/html')
+				.documentElement.textContent;
+		}
+	}
+});
+
+/**
+ * Returns an HTML string converted to plain text, with all HTML tags removed.
+ */
+String.prototype.toPlainText || Object.defineProperties(String.prototype, {
+	toPlainText: {
+		value: function () { // preserves `this`
+			return this
+				// treat new lines and subsequent indents as white spaces
+				.replace(/(\r\n|\n|\r)[ \t]*/g, ' ')
+				// remove special HTML blocks
+				.replace(/<(audio|canvas|noscript|script|style|video)(\s|>).*?<\/\1>/gi, '')
+				// convert block-level HTML tags to new lines
+				.replace(/<\/?(address|article|aside|blockquote|dd|div|dl|dt|fieldset|figcaption|figure|footer|form|h[1-6]|header|li|main|nav|ol|p|pre|section|table|tfoot|ul)(\s[^>]*>|>)/gi, '\n')
+				// convert special HTML tags to new lines
+				.replace(/<(br|hr)\s?\/?>/gi, '\n')
+				// remove remaining HTML tags
+				.replace(/<[^>]+>/g, '')
+				.replaceAll('  ', ' ')
+				.trimAll()
+				.htmlDecode();
 		}
 	}
 });