+++ CHANGELOG.md
@@ -1,11 +1,22 @@
 # Changes to Balzac.js
 
+### 0.0.6 (July 25, 2023)
+
+* Faster library initialization.
+* Improve String.toTitleCase and String.toSentenceCase functions.
+* Add String.trimAll and String.htmlDecode functions.
+* String.stripTags function replaced by enhanced String.toPlainText.
+* Remove document.DOMContentLoaded function.
+* Improve HTML elements case attribute.
+* Improve HTML form elements validation.
+* Update README.md and test.html files.
+
 ### 0.0.5 (July 22, 2023)
 
 * Add String.isUpperCase and String.isLowerCase functions.
 * Improve String.toTitleCase and String.toSentenceCase functions.
 * Rename String.encodeHtml function to htmlEncode.
-* Contents of HTML output elements cleared when resetting form.
+* Content of HTML output elements cleared when resetting form.
 
 ### 0.0.4 (July 22, 2023)
 
@@ -22,7 +33,7 @@
 * Fix RegExp.PATTERN.EMAIL, RegExp.PATTERN.EMAILS and RegExp.PATTERN.PHONE expressions.
 * Add RegExp.PATTERN.IBAN expression and related "iban" HTML input element.
 * Add RegExp.PATTERN.POSTCODE expression and related "postcode" HTML input element.
-* Remove whitespaces from both ends of text input values when case="upper" or case="lower".
+* Remove white spaces from both ends of text input values when case="upper" or case="lower".
 * Rename HTMLInputElement.PASSWORD_RATING.FAIR constant to MEDIUM.
 * Update README.md and test.html files.